+++ core/modules/cart/store/actions/connectActions.ts
@@ -9,28 +9,20 @@ const connectActions = {
   toggleMicrocart ({ commit }) {
     commit(types.CART_TOGGLE_MICROCART)
   },
-  async clear ({ commit, dispatch, getters }, options = { recreateAndSyncCart: true }) {
+  async clear ({ commit, dispatch, getters }) {
     await commit(types.CART_LOAD_CART, [])
-    if (options.recreateAndSyncCart && getters.isCartSyncEnabled) {
-      await commit(types.CART_LOAD_CART_SERVER_TOKEN, null)
-      await commit(types.CART_SET_ITEMS_HASH, null)
-      await dispatch('connect', { guestCart: !config.orders.directBackendSync }) // guest cart when not using directBackendSync because when the order hasn't been passed to Magento yet it will repopulate your cart
-    }
+    await commit(types.CART_LOAD_CART_SERVER_TOKEN, null)
+    await commit(types.CART_SET_ITEMS_HASH, null)
   },
   async disconnect ({ commit }) {
     commit(types.CART_LOAD_CART_SERVER_TOKEN, null)
   },
   async authorize ({ dispatch, getters }) {
+    await dispatch('connect', { guestCart: false })
+
     const coupon = getters.getCoupon.code
-    const lastCartBypassTs = await StorageManager.get('user').getItem('last-cart-bypass-ts')
-    const timeBypassCart = config.orders.directBackendSync || (Date.now() - lastCartBypassTs) >= (1000 * 60 * 24)
-
-    if (!config.cart.bypassCartLoaderForAuthorizedUsers || timeBypassCart) {
-      await dispatch('connect', { guestCart: false })
-
-      if (!getters.getCoupon) {
-        await dispatch('applyCoupon', coupon)
-      }
+    if (!getters.getCoupon) {
+      await dispatch('applyCoupon', coupon)
     }
   },
   async connect ({ getters, dispatch, commit }, { guestCart = false, forceClientState = false }) {
@@ -53,6 +45,17 @@ const connectActions = {
 
     Logger.warn('Cart sync is disabled by the config', 'cart')()
     return createDiffLog()
+  },
+  /**
+   * Create cart token when there are products in cart and we don't have token already
+   */
+  async create ({ dispatch, getters }) {
+    const storedItems = getters['getCartItems'] || []
+    const cartToken = getters['getCartToken']
+    if (storedItems.length && !cartToken) {
+      Logger.info('Creating server cart token', 'cart')()
+      await dispatch('connect', { guestCart: false })
+    }
   }
 }